requirements.txt
streamlit
PyMuPDF
openai>=1.0.0
python-dotenv
